--- src/main/java/com/image/manager/edgeserver/controllers/TestController.java
@@ -1,16 +1,0 @@
-package com.image.manager.edgeserver.controllers;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/test")
-public class TestController {
-
-    @GetMapping
-    public String hello() {
-        return "Hello World!";
-    }
-
-}
