+++ src/main/java/com/image/manager/edgeserver/domain/operation/CropOperation.java
@@ -1,4 +1,4 @@
-package com.image.manager.edgeserver.model;
+package com.image.manager.edgeserver.domain.operation;
 
 import lombok.EqualsAndHashCode;
