+++ src/main/java/com/image/manager/edgeserver/domain/operation/Operation.java
@@ -1,4 +1,4 @@
-package com.image.manager.edgeserver.model;
+package com.image.manager.edgeserver.domain.operation;
 
 import java.awt.image.BufferedImage;
