+++ src/main/java/com/image/manager/edgeserver/domain/operation/OperationFactory.java
@@ -1,4 +1,4 @@
-package com.image.manager.edgeserver.model;
+package com.image.manager.edgeserver.domain.operation;
 
 import org.springframework.stereotype.Component;
