+++ src/main/java/com/image/manager/edgeserver/domain/operation/OperationType.java
@@ -1,0 +1,6 @@
+package com.image.manager.edgeserver.domain.operation;
+
+public enum OperationType {
+    SCALE,
+    CROP
+}
