--- src/main/java/com/image/manager/edgeserver/model/OperationType.java
@@ -1,6 +1,0 @@
-package com.image.manager.edgeserver.model;
-
-public enum OperationType {
-    SCALE,
-    CROP
-}
