+++ src/test/java/com/image/manager/edgeserver/EdgeServerApplicationTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class EdgeServerApplicationTests {
+class EdgeServerApplicationTest {
 
     @Test
     void contextLoads() {
