+++ src/test/java/com/image/manager/edgeserver/domain/operation/OperationFactoryTest.java
@@ -1,10 +1,12 @@
-package com.image.manager.edgeserver.model;
+package com.image.manager.edgeserver.domain.operation;
 
+import com.image.manager.edgeserver.domain.operation.CropOperation;
+import com.image.manager.edgeserver.domain.operation.Operation;
+import com.image.manager.edgeserver.domain.operation.OperationFactory;
+import com.image.manager.edgeserver.domain.operation.ScaleOperation;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 import java.util.List;
